vmm: Use set_immediate_exit to avoid a race on pausing the vCPU
A race can occur if the signal to the vCPU thread is sent between the
shared state atomic being checked (e.g. for paused) and entering
KVM_RUN. As such we can use KVM's immediate_exit boolean to make the
KVM_RUN exit immediately.

We do this in twice - the first time we only try and take the lock -
this would prevent the pause being unduly waiting for the KVM_RUN ioctl
to exit (the most likely reason for the lock being held).  This is the
main solution to the race.

We also do it again after signalling - this is to
handle the rare case where the lock is held for some other than the
wrapper around KVM_RUN - in this case the majority of the time the lock
won't block as the vCPU would already have been kicked out of KVM_RUN by
the signal. However this ensures that we don't re-enter KVM_RUN ioctl.

Fixes: #7427

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>